+++ epidatpy/__init__.py
@@ -1,27 +1,9 @@
 """Fetch data from Delphi's API."""
 
-from ._constants import __version__
-from ._model import (
-    EpiRange,
-    EpiRangeDict,
-    EpiDataResponse,
-    EpiRangeLike,
-    InvalidArgumentException,
-    EpiRangeParam,
-    IntParam,
-    StringParam,
-    EpiDataFormatType,
-    AEpiDataCall,
-)
-from ._covidcast import (
-    DataSignal,
-    DataSource,
-    WebLink,
-    DataSignalGeoStatistics,
-    CovidcastDataSources,
-    GeoType,
-    TimeType,
-)
-from ._auth import get_api_key
+# Make the linter happy about the unused variables
+__all__ = ["__version__", "Epidata", "CovidcastEpidata", "EpiRange"]
+__author__ = "Delphi Research Group"
 
-__author__ = "Delphi Group"
+
+from ._constants import __version__
+from .request import CovidcastEpidata, Epidata, EpiRange
